make "%llu" less likely to warn on weird platforms
svn:r2209

+++ trunk/src/or/main.c
@@ -965,7 +965,7 @@ static void dumpstats(int severity) {
 
   if (stats_n_seconds_uptime)
     log(severity,"Average bandwidth used: %llu/%ld = %d bytes/sec",
-           stats_n_bytes_read, stats_n_seconds_uptime,
+           (unsigned long long) stats_n_bytes_read, stats_n_seconds_uptime,
            (int) (stats_n_bytes_read/stats_n_seconds_uptime));
 
   rep_hist_dump_stats(now,severity);
